feat: [feature/wishlist-back] wishlist 관련  api 구현

+++ server/src/main/java/com/kofoos/api/wishlist/WishlistRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface WishlistRepository extends JpaRepository<WishlistItem, Long> {
+public interface WishlistRepository extends JpaRepository<WishlistItem, Integer> {
 
     Optional<WishlistItem> findWishlistItemByWishlistFolderIdAndProductId(int wishlist_folder_id, int productId);
